v1.0.0 :: redvsblue.money app
Includes the contract ABI to instatiate connections to Kovan (42)
and Mainnet (1). Game is functional and tested locally and on said
test net. All tests pass etc. No fee model verified, along with deposit
and withdrawal.

See the `Makefile` for deployment rules if cloning this locally.

Added a tip link to the README and the app.